requirements.txt
numpy==1.19.5
streamlit==0.86.0
tensorflow==2.6.0
Pillow==8.3.1
